requirements.txt
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.5
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.62.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==26.0
pillow==12.0.0
protobuf==7.34.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
triton==3.3.1
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
